+++ client/index.ts
@@ -1,0 +1,2 @@
+export { ChatClient } from "./chat-client";
+export { Client } from "./client";
