--- constants.ts
@@ -1,1 +1,0 @@
-export const rootApiUrl = "https://api.commonbase.com";
